impl value executor2
